refactor(search): rename package
