+++ lib/event/events_screen.dart
@@ -6,11 +6,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
-import 'package:x_qrcode/attendees/attendees_screen.dart';
+import 'package:x_qrcode/attendee/attendees_screen.dart';
 import 'package:x_qrcode/common/app_drawer.dart';
-import 'package:x_qrcode/common/common_models.dart';
-import 'package:x_qrcode/organization/user.dart';
-import 'package:x_qrcode/visitors/visitors_screen.dart';
+import 'package:x_qrcode/common/constants.dart';
+import 'package:x_qrcode/organization/model/user_model.dart';
+import 'package:x_qrcode/visitor/visitors_screen.dart';
 
 import '../constants.dart';
