+++ lib/visitor/consent_screen.dart
@@ -7,11 +7,11 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
 import 'package:signature/signature.dart';
 import 'package:x_qrcode/api/api_service.dart';
-import 'package:x_qrcode/events/events_screen.dart';
-import 'package:x_qrcode/organization/user.dart';
-import 'package:x_qrcode/visitors/attendee.dart';
+import 'package:x_qrcode/event/events_screen.dart';
+import 'package:x_qrcode/organization/model/user_model.dart';
 
 import '../constants.dart';
+import 'model/attendee_model.dart';
 
 class ConsentScreenArguments {
   final String visitorId;
