+++ lib/visitor/model/attendee_model.dart
